combine const.py with constants.py
